+++ CHANGELOG.md
@@ -5,17 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.4.0]
+
+### Added
+
+- Ripp class partially implemented. Allows to add RiPP-specific elements. Input functions for "peptidases" and "precursor genes" are still missing
+
+### Changed
+
+- Changed structure of main() to accommodate optional input data
+- Changed json schema to allow the entry "mod_precursor_pep". Allows to add the modified precursor peptide.
+
+
 ## [0.3.0] 12-06-2023
 
 ### Changed
 
 - Renamed class Base to MibigEntry.
-- The MibigEntry class now generates a dict
-    called mibig_entry that is consecutively modified by other classes.
+- The MibigEntry class now generates a dict called mibig_entry that is consecutively modified by other classes.
 - Class Changelog now modifies the mibig_entry dict instead of generating a new dict.
 - Class WriteMibig now uses the mibig_entry dict directly instead of concatenating
     the dicts into a new dict.
-
 
 ## [0.2.0] 11-06-2023